fix & test ID example

+++ src/Regula.FaceSDK.WebClient/Model/Group.cs
@@ -37,7 +37,7 @@ namespace Regula.FaceSDK.WebClient.Model
         /// <param name="metadata">metadata.</param>
         /// <param name="id">id.</param>
         /// <param name="createdAt">createdAt.</param>
-        public Group(string name = default(string), Dictionary<string, Object> metadata = default(Dictionary<string, Object>), decimal id = default(decimal), DateTime createdAt = default(DateTime))
+        public Group(string name = default(string), Dictionary<string, Object> metadata = default(Dictionary<string, Object>), int id = default(int), DateTime createdAt = default(DateTime))
         {
             this.Name = name;
             this.Metadata = metadata;
@@ -61,7 +61,7 @@ namespace Regula.FaceSDK.WebClient.Model
         /// Gets or Sets Id
         /// </summary>
         [DataMember(Name="id", EmitDefaultValue=false)]
-        public decimal Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Gets or Sets CreatedAt
