+++ src/Regula.FaceSDK.WebClient/Model/SearchResult.cs
@@ -40,7 +40,7 @@ namespace Regula.FaceSDK.WebClient.Model
         /// <param name="updatedAt">updatedAt.</param>
         /// <param name="images">images.</param>
         /// <param name="code">code.</param>
-        public SearchResult(string name = default(string), Dictionary<string, Object> metadata = default(Dictionary<string, Object>), decimal id = default(decimal), DateTime createdAt = default(DateTime), DateTime updatedAt = default(DateTime), List<RecognizeImage> images = default(List<RecognizeImage>), decimal code = default(decimal))
+        public SearchResult(string name = default(string), Dictionary<string, Object> metadata = default(Dictionary<string, Object>), int id = default(int), DateTime createdAt = default(DateTime), DateTime updatedAt = default(DateTime), List<RecognizeImage> images = default(List<RecognizeImage>), int code = default(int))
         {
             this.Name = name;
             this.Metadata = metadata;
@@ -67,7 +67,7 @@ namespace Regula.FaceSDK.WebClient.Model
         /// Gets or Sets Id
         /// </summary>
         [DataMember(Name="id", EmitDefaultValue=false)]
-        public decimal Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Gets or Sets CreatedAt
@@ -91,7 +91,7 @@ namespace Regula.FaceSDK.WebClient.Model
         /// Gets or Sets Code
         /// </summary>
         [DataMember(Name="code", EmitDefaultValue=false)]
-        public decimal Code { get; set; }
+        public int Code { get; set; }
 
         /// <summary>
         /// Returns the string presentation of the object
